+++ src/macros.rs
@@ -22,10 +22,10 @@
 /// fn callback2() -> Option<usize> { 2_usize.checked_mul(2) }
 ///
 /// benches!(
-///     Bench::new("usize", "checked_add(2)")
-///         .with(callback),
-///     Bench::new("usize", "checked_mul(2)")
-///         .with(callback2)
+///     Bench::new("usize::checked_add(2)")
+///         .run(callback),
+///     Bench::new("usize::checked_mul(2)")
+///         .run(callback2)
 /// );
 /// ```
 macro_rules! benches {
@@ -33,15 +33,16 @@ macro_rules! benches {
 		/// # Benchmarks!
 		fn main() {
 			// This can take a while; give 'em a message of hope.
-			::std::print!("\x1b[1;38;5;199mStarting:\x1b[0m Running benchmark(s). Stand by!\n\n");
+			::std::eprint!("\x1b[1;38;5;199mStarting:\x1b[0m Running benchmark(s). Stand by!\n\n");
 
 			// Run the benches.
-			let mut benches: Vec<$crate::Bench> = Vec::new();
+			let mut benches = $crate::Benches::default();
 			$(
+				::std::eprint!("\x1b[1;38;5;4m•\x1b[0m");
 				benches.push($benches);
 			)+
-
-			$crate::analyze(&mut benches);
+			::std::eprintln!("\n");
+			benches.finish();
 		}
 	};
 }